refactor: Separate utils from handlers
Also improves bot error message
